backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.22.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.8
python-dotenv==1.0.0
openai==1.0.0
pydantic==1.10.11
jinja2==3.1.2
langchain==0.1.0
langchain-openai==0.0.5
python-dateutil==2.8.2